requirements.txt
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-bootstrap5
